Implement the empty_values options
Add keywords to the __init__() method to specify empty_values or allowed_empty_values to specify what should or should not be considered an empty/blank value.  This happens for both the Form field and the database field.  The database field will in turn instantiate the Form field appropriately.

With these changes, the tests pass and developers should be able to specify which JSON values should be considered blank/null, and which should not.